Add screen to choose type when creating new wallet
