+++ app/src/main/java/io/horizontalsystems/bankwallet/uiv3/components/Section.kt
@@ -1,0 +1,23 @@
+package io.horizontalsystems.bankwallet.uiv3.components
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.unit.dp
+import io.horizontalsystems.bankwallet.ui.compose.ComposeAppTheme
+
+@Composable
+fun Section(content: @Composable () -> Unit) {
+    Column(
+        modifier = Modifier
+            .padding(horizontal = 16.dp)
+            .clip(RoundedCornerShape(16.dp))
+            .background(ComposeAppTheme.colors.lawrence)
+    ) {
+        content()
+    }
+}
