+++ app/src/main/res/navigation/main_graph.xml
@@ -244,6 +244,9 @@
         android:id="@+id/createAccountFragment"
         android:name="io.horizontalsystems.bankwallet.modules.createaccount.CreateAccountFragment" />
     <fragment
+        android:id="@+id/createAccountStandardFragment"
+        android:name="io.horizontalsystems.bankwallet.modules.createaccount.CreateAccountStandardFragment" />
+    <fragment
         android:id="@+id/restoreAccountFragment"
         android:name="io.horizontalsystems.bankwallet.modules.restoreaccount.RestoreAccountFragment" />
     <fragment
